ci(move_files): move stragglers into organized folders
