+++ test/fast/commands/branch/branch_checkout.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
-import { allScenes } from "../../lib/scenes";
-import { configureTest } from "../../lib/utils";
+import { allScenes } from "../../../lib/scenes";
+import { configureTest } from "../../../lib/utils";
 
 for (const scene of allScenes) {
   describe(`(${scene}): branch create`, function () {
